+++ dialect/src/test/java/ydb/jimmer/dialect/YdbTest.java
@@ -9,7 +9,6 @@ import org.springframework.core.io.support.EncodedResource;
 import org.springframework.jdbc.datasource.init.ScriptException;
 import org.springframework.jdbc.datasource.init.ScriptUtils;
 import tech.ydb.test.junit5.YdbHelperExtension;
-import ydb.jimmer.client.YdbDialect;
 import ydb.jimmer.dialect.sqlMonitor.ExecutorMonitor;
 
 import java.net.URL;
